refactor tests for projects django app
