--- apps/projects/testing/__init__.py
@@ -1,1 +1,0 @@
-from .tests import *
